+++ 寿司チョコレート-流れ魚/寿司チョコレート-流れ魚/SceneMain.h
@@ -17,5 +17,5 @@ private:
 	int m_time;
 	int x;
 	int t = 0;
-	int sp = 5;
+	float sp = 5.0f;
 };
